Review syncCommittee validation
