+++ packages/lodestar/src/chain/validation/signatureSets/index.ts
@@ -1,0 +1,6 @@
+export * from "./aggregateAndProof";
+export * from "./contributionAndProof";
+export * from "./selectionProof";
+export * from "./syncCommittee";
+export * from "./syncCommitteeContribution";
+export * from "./syncCommitteeSelectionProof";
