+++ packages/lodestar/src/chain/validation/syncCommittee.ts
@@ -1,7 +1,6 @@
 import {
   CachedBeaconState,
   computeSubnetsForSyncCommittee,
-  fast,
   getIndicesInSubSyncCommittee,
 } from "@chainsafe/lodestar-beacon-state-transition";
 import {IBeaconConfig} from "@chainsafe/lodestar-config";
@@ -10,7 +9,11 @@ import {readonlyValues} from "@chainsafe/ssz";
 import {IBeaconDb} from "../../db";
 import {ISyncCommitteeJob, SyncCommitteeError, SyncCommitteeErrorCode} from "../errors/syncCommitteeError";
 import {IBeaconChain} from "../interface";
+import {getSyncCommitteeSignatureSet} from "./signatureSets";
 
+/**
+ * Spec v1.1.0-alpha.3
+ */
 export async function validateGossipSyncCommittee(
   config: IBeaconConfig,
   chain: IBeaconChain,
@@ -19,7 +22,7 @@ export async function validateGossipSyncCommittee(
   subnet: number
 ): Promise<void> {
   const {syncCommittee, validSignature} = syncCommitteeJob;
-  // The signature's slot is for the current slot
+  // [IGNORE] The signature's slot is for the current slot, i.e. sync_committee_signature.slot == current_slot.
   if (chain.clock.currentSlot < syncCommittee.slot) {
     throw new SyncCommitteeError({
       code: SyncCommitteeErrorCode.FUTURE_SLOT,
@@ -28,7 +31,6 @@ export async function validateGossipSyncCommittee(
       job: syncCommitteeJob,
     });
   }
-
   if (chain.clock.currentSlot > syncCommittee.slot) {
     throw new SyncCommitteeError({
       code: SyncCommitteeErrorCode.PAST_SLOT,
@@ -38,7 +40,7 @@ export async function validateGossipSyncCommittee(
     });
   }
 
-  // The block being signed over has not been seen
+  // [IGNORE] The block being signed over (sync_committee_signature.beacon_block_root) has been seen (via both gossip and non-gossip sources).
   if (!chain.forkChoice.hasBlock(syncCommittee.beaconBlockRoot)) {
     throw new SyncCommitteeError({
       code: SyncCommitteeErrorCode.UNKNOWN_BEACON_BLOCK_ROOT,
@@ -47,7 +49,8 @@ export async function validateGossipSyncCommittee(
     });
   }
 
-  // There has been no other valid sync committee signature with same slot and validatorIndex
+  // [IGNORE] There has been no other valid sync committee signature for the declared slot for the validator referenced
+  // by sync_committee_signature.validator_index.
   if (db.seenSyncCommiteeCache.hasSyncCommitteeSignature(subnet, syncCommittee)) {
     throw new SyncCommitteeError({
       code: SyncCommitteeErrorCode.SYNC_COMMITTEE_ALREADY_KNOWN,
@@ -55,8 +58,9 @@ export async function validateGossipSyncCommittee(
     });
   }
 
+  // [REJECT] The subnet_id is valid for the given validator, i.e. subnet_id in compute_subnets_for_sync_committee(state, sync_committee_signature.validator_index).
+  // Note this validation implies the validator is part of the broader current sync committee along with the correct subcommittee.
   const headState = chain.getHeadState();
-  // validate validator being part of the current sync committee
   const validatorPubkey = headState.validators[syncCommittee.validatorIndex].pubkey;
   const currentSyncCommittee = (headState as CachedBeaconState<altair.BeaconState>).currentSyncCommittee;
   const syncCommitteePubkeys = Array.from(readonlyValues(currentSyncCommittee.pubkeys));
@@ -70,6 +74,7 @@ export async function validateGossipSyncCommittee(
     });
   }
 
+  // TODO: This check validates the check above implicitly
   // validate subnet
   const expectedSubnets = computeSubnetsForSyncCommittee(
     config,
@@ -85,9 +90,9 @@ export async function validateGossipSyncCommittee(
     });
   }
 
-  // validate signature
   if (!validSignature) {
-    const signatureSet = fast.getSyncCommitteeSignatureSet(headState, syncCommittee);
+    // [REJECT] The signature is valid for the message beacon_block_root for the validator referenced by validator_index.
+    const signatureSet = getSyncCommitteeSignatureSet(headState, syncCommittee);
     if (!(await chain.bls.verifySignatureSets([signatureSet]))) {
       throw new SyncCommitteeError({
         code: SyncCommitteeErrorCode.INVALID_SIGNATURE,
